refactor(jira): Extracted jira creation into a service class
- created jiraService class containing jira creation bits
